Add --remove-logs flag to pki-destroy
Partially resolves: Bug 1372056

List of changes by this commit:

- Logs are preserved by default (comment #1 in BZ)
- Add `--remove-flags` flag to pkidestroy to remove logs

+++ base/server/python/pki/server/deployment/scriptlets/instance_layout.py
@@ -199,8 +199,11 @@ class PkiScriptlet(pkiscriptlet.AbstractBasePkiScriptlet):
 
         # remove Tomcat instance base
         deployer.directory.delete(deployer.mdict['pki_instance_path'])
-        # remove Tomcat instance logs
-        deployer.directory.delete(deployer.mdict['pki_instance_log_path'])
+
+        # remove Tomcat instance logs only if --remove-logs is specified
+        if deployer.mdict['pki_remove_logs']:
+            deployer.directory.delete(deployer.mdict['pki_instance_log_path'])
+
         # remove shared NSS security database path for this instance
         deployer.directory.delete(deployer.mdict['pki_database_path'])
         # remove Tomcat instance configuration
